cast tv_usec to long for *printf
on darwin suseconds_t is an int.

ok tb@

+++ usr.sbin/smtpd/lka_filter.c
@@ -925,13 +925,13 @@ filter_protocol_query(struct filter *filter, uint64_t token, uint64_t reqid, con
 		n = io_printf(lka_proc_get_io(filter->proc),
 		    "filter|%s|%lld.%06ld|smtp-in|%s|%016"PRIx64"|%016"PRIx64"|%s|%s\n",
 		    PROTOCOL_VERSION,
-		    (long long)tv.tv_sec, tv.tv_usec,
+		    (long long)tv.tv_sec, (long)tv.tv_usec,
 		    phase, reqid, token, fs->rdns, param);
 	else
 		n = io_printf(lka_proc_get_io(filter->proc),
 		    "filter|%s|%lld.%06ld|smtp-in|%s|%016"PRIx64"|%016"PRIx64"|%s\n",
 		    PROTOCOL_VERSION,
-		    (long long)tv.tv_sec, tv.tv_usec,
+		    (long long)tv.tv_sec, (long)tv.tv_usec,
 		    phase, reqid, token, param);
 	if (n == -1)
 		fatalx("failed to write to processor");
@@ -949,7 +949,7 @@ filter_data_query(struct filter *filter, uint64_t token, uint64_t reqid, const c
 	    "filter|%s|%lld.%06ld|smtp-in|data-line|"
 	    "%016"PRIx64"|%016"PRIx64"|%s\n",
 	    PROTOCOL_VERSION,
-	    (long long)tv.tv_sec, tv.tv_usec,
+	    (long long)tv.tv_sec, (long)tv.tv_usec,
 	    reqid, token, line);
 	if (n == -1)
 		fatalx("failed to write to processor");
@@ -1366,7 +1366,7 @@ report_smtp_broadcast(uint64_t reqid, const char *direction, struct timeval *tv,
 		va_start(ap, format);
 		if (io_printf(lka_proc_get_io(rp->name),
 		    "report|%s|%lld.%06ld|%s|%s|%016"PRIx64"%s",
-		    PROTOCOL_VERSION, (long long)tv->tv_sec, tv->tv_usec,
+		    PROTOCOL_VERSION, (long long)tv->tv_sec, (long)tv->tv_usec,
 		    direction, event, reqid,
 		    format[0] != '\n' ? "|" : "") == -1 ||
 		    io_vprintf(lka_proc_get_io(rp->name), format, ap) == -1)
